Complete parallel implementation of factorial
This is implemented with the raylon framework. For smaller numbers like
100 the serial implementation is still significantly faster, but when
you talk about numbers like 1000000 factorial, the parallel
implementation is miles faster than it's serial counterpart. 1000000!
goes from about 44 seconds in it's serial implementation to 3 seconds in
the parallel implementation.

Implementation:
Work is split into chunks whose size is based on the number of threads
available in the thread pool. The factorial of these chunks is then
calculated on different threads and the result of this work is
multiplied together on the main thread.